example/terrafrom/src/requirements.txt
glados
slackclient==2.3.1